# Reduce allocations in `pem::encode`